+++ kdal/src/main/java/com/weather/android/kdal/model/V2idxDriveCurrent.kt
@@ -6,7 +6,7 @@ data class V2idxDriveCurrent(
         val drivingDifficultyIndexCurrent: DrivingDifficultyIndexCurrent
 ) {
     data class DrivingDifficultyIndexCurrent(
-            val obsTime: Int, // 1529678934
+            val obsTime: Long, // 1529678934
             val obsTimeLocal: String, // 2018-06-22T09:48:54-0500
             val dow: String, // Friday
             val dayInd: String, // D
@@ -20,7 +20,7 @@ data class V2idxDriveCurrent(
             val version: String, // 1
             val latitude: Double, // 29.19
             val longitude: Double, // -96.27
-            val expireTimeGmt: Int, // 1529679584
+            val expireTimeGmt: Long, // 1529679584
             val statusCode: Int // 200
     )
 }
